swagger working again with serverConnect.js
