+++ README.md
@@ -24,10 +24,19 @@ npm --mongo_db_for_testing=host_address test
 npm --mongo_db_for_testing=host_address --web_port=port_number test
 ```
 ### Swagger Documentation
-Before running, configure the db.js file to mongodb and the corresponding database.
-queries.js points to port 8080 to retrieve the endpointsfrom server.js, this may be changed depending on where server.js is running.
+queries.js points to port 8080 to retrieve the endpoints from server.js, this may be changed depending on where server.js is running.
 To run the documentation (assuming set up was complete):
 ```
-swagger project start
+host=host_address database=database_name swagger project start
 ```
-A link will pop up. go to that link and add /docs to the end.
+as environment variables, you can also add PORT for where swagger will end up, and web_port for the api port. (if changing web_port, queries.js must be updated too)
+A link will pop up in the command line. go to that link and add /docs to the end.
+
+The expected output to the console:
+```
+Starting: /home/<name>/seqware-browser-REST/serverConnect.js...
+  project started here: http://localhost:10010/
+  project will restart on changes.
+  to restart at any time, enter `rs`
+Magic happens on port 8080
+```
