+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "private": false,
   "description": "",
-  "main": "server.js",
+  "main": "serverConnect.js",
   "scripts": {
     "test": "./node_modules/.bin/mocha --reporter spec",
     "start": "node serverConnect.js"
